Added CONFIG library code.
Added the template file QQQ_CONFIG.DSPIC, and the library include file
CONFIG.INS.DSPIC.  The config module allows applications to get static
config data by ID.  The structure of how the config data is stored in
non-volatile memory is not surfaced to applications.  It is documented in
the CONFIG.INS.DSPIC library file.