Made CMD and CMDS modules generic.
The CMD (command processor) and CMDS (various command routines) modules
have now been made generic.  Each has a QQQ_ file that includes the bulk
of the code from a "library" file that is not intended to be copied into
each project.

Each implemented command is now specified at a higher level in each
project.  The new CMD and CMDS code configures itself automatically to
which commands are implemented.  This new mechanism allows for various
generic commands to be written, but then only included in a project if
actually referenced.  Unreferenced command routines cause no code to be
added to a project.

It is intended that other subsystems that commonly have commands
associated with them now use this mechanism to provide standard versions
of those commands.  These produce no code unless specifically referenced.