Expanded standard FWINFO command.
The FWINFO command was previously implemented in the CMD module, and only
sent the info for its firmware.  The FWINFO command and SEND_FWINFO
subroutine are now in the CMDS module.  This routine now also reports the
USB device firmware info, if the USTAT command is enabled and the USB
device firmware is known.