SYS module in test mode code from 254 to 1.
The CFGSTAT response was changed so that the status code for the SYS
module for being in test mode is now 1, was 254.  This make it compatible
with the existing CANSER STAT response, which was implemented before this
generic library module.  The CANSER firmware now uses this module instead
of its own code for sending its STAT response.