Moved DYMEM_HEAP definition to STD.INS.DSPIC.
The DYMEM_HEAP constant indicates whether the Embed DYMEM heap is in use.
This is now guaranteed to exist by STD.INS.DSPIC, and can therefore be
relied on to exist by any other module.  Currently, DYMEM_HEAP	is set to
TRUE if the constant MINSTACK0 exists.  MINSTACK0 was not previously used,
but is required by the DYMEM routines for configuring the heap.