Update how MDEVs get configured.
There was previously explicit code in the INIT module to configure SYS,
then run the configuration routines of all remaining MDEVs.  This has now
been moved to the CONFIG_MDEVS subroutine in the CONFIG module.

Also fixed bug where failure flag was not correctly set when a
configuration status with the high bit set was encountered.